fall back1

+++ python/sglang/srt/models/glm4_moe.py
@@ -329,10 +329,7 @@ class Glm4MoeGate(nn.Module):
         )
 
     def forward(self, hidden_states):
-        # Cast to FP32 before gate projection for GLM-V model.
-        logits = F.linear(
-            hidden_states.to(torch.float32), self.weight.to(torch.float32), None
-        )
+        logits = F.linear(hidden_states, self.weight, None)
         return logits
